breadth, categories, and weights added to scraper
