+++ cat_data.csv
@@ -1,0 +1,263 @@
+,category,breadth
+0,ACTURSCI,['C']
+1,ASL,['B']
+2,AMERICAN,['A']
+3,ADS,['A']
+4,ANATCELL,['C']
+5,ANTHRO,['A']
+6,APPLMATH,['C']
+7,ARABIC,['B']
+8,AH,['B']
+9,AISE,['C']
+10,ARTHUM,['B']
+11,ASTRONOM,['C']
+12,BIOCHEM,['C']
+13,BIOLOGY,['C']
+14,BME,['C']
+15,BIOSTATS,['C']
+16,BUSINESS,['A']
+17,CALCULUS,['C']
+18,CANADIAN,['A']
+19,CGS,['A']
+20,CHEMBIO,['C']
+21,CBE,['C']
+22,CHEM,['C']
+23,CYS,['A']
+24,CHINESE,['B']
+25,CEE,['C']
+26,CLASSICS,['B']
+27,COMMSCI,['C']
+28,COMPLIT,['B']
+29,COMPSCI,['C']
+30,CA,"['A', 'B']"
+31,DANCE,['A']
+32,DATASCI,['C']
+33,DIGICOMM,['A']
+34,DIGIHUM,['B']
+35,DISABST,['A']
+36,EARTHSCI,['C']
+37,ECONOMIC,['A']
+38,EDUC,['A']
+39,ECE,['C']
+40,ELI,['C']
+41,ENGSCI,['C']
+42,ENGLISH,['B']
+43,ENVIRSCI,['C']
+44,EPID,['C']
+45,EPIDEMIO,['C']
+46,FIMS,['A']
+47,FAMLYSTU,['A']
+48,FILM,['B']
+49,FINMOD,['C']
+50,FOODNUTR,['A']
+51,FRENCH,['B']
+52,GSWS,"['A', 'B']"
+53,GEOGRAPH,['A']
+54,GERMAN,['B']
+55,GGB,['B']
+56,GLE,['A']
+57,GREEK,['B']
+58,GPE,['C']
+59,HEALTSCI,['A']
+60,HEBREW,['B']
+61,HINDI,['B']
+62,HISTORY,['A']
+63,HISTSCI,['C']
+64,HUMANECO,['A']
+65,HUMANRS,['A']
+66,HUMANIT,[]
+67,INDIGSTU,['A']
+68,IE,['C']
+69,INTEGSCI,['C']
+70,ICC,['B']
+71,INTERDIS,['A']
+72,INTREL,['A']
+73,ITALIAN,['B']
+74,ITALST,['B']
+75,JAPANESE,['B']
+76,JEWISH,['A']
+77,KINESIOL,['A']
+78,KOREAN,['B']
+79,LATIN,['B']
+80,LAW,['A']
+81,LS,['A']
+82,LINGUIST,"['A', 'B']"
+83,MOS,['A']
+84,MATH,['C']
+85,MME,['C']
+86,MSE,['C']
+87,MIT,['A']
+88,MBI,['C']
+89,MEDBIO,['C']
+90,MEDSCIEN,['C']
+91,MEDIEVAL,['B']
+92,MICROIMM,['C']
+93,MCS,['B']
+94,MUSIC,['A']
+95,NEURO,['C']
+96,NMM,['C']
+97,NURSING,['A']
+98,ONEHEALT,['C']
+99,PATHOL,['C']
+100,PERSIAN,['B']
+101,PHARM,['C']
+102,PHILOSOP,['B']
+103,PHYSICS,['C']
+104,PHYSIOL,['C']
+105,PHYSPHRM,['C']
+106,POLISCI,['A']
+107,PPE,['A']
+108,PORTUGSE,['B']
+109,PSYCHOL,['A']
+110,REHABSCI,['A']
+111,RELSTUD,['B']
+112,RUSSIAN,['B']
+113,SCHOLARS,[]
+114,SCIENCE,['C']
+115,SOCLJUST,['A']
+116,SOCSCI,['A']
+117,SOCWORK,['A']
+118,SOCIOLOG,['A']
+119,SE,['C']
+120,SPANISH,['B']
+121,SPEECH,['B']
+122,STATS,['C']
+123,SA,['B']
+124,THANAT,['A']
+125,TNLA,['B']
+126,THEATRE,['B']
+127,THEOLST,['B']
+128,TJ,['A']
+129,WORLDLIT,['B']
+130,WRITING,['B']
+131,ACTURSCI,['C']
+132,ASL,['B']
+133,AMERICAN,['A']
+134,ADS,['A']
+135,ANATCELL,['C']
+136,ANTHRO,['A']
+137,APPLMATH,['C']
+138,ARABIC,['B']
+139,AH,['B']
+140,AISE,['C']
+141,ARTHUM,['B']
+142,ASTRONOM,['C']
+143,BIOCHEM,['C']
+144,BIOLOGY,['C']
+145,BME,['C']
+146,BIOSTATS,['C']
+147,BUSINESS,['A']
+148,CALCULUS,['C']
+149,CANADIAN,['A']
+150,CGS,['A']
+151,CHEMBIO,['C']
+152,CBE,['C']
+153,CHEM,['C']
+154,CYS,['A']
+155,CHINESE,['B']
+156,CEE,['C']
+157,CLASSICS,['B']
+158,COMMSCI,['C']
+159,COMPLIT,['B']
+160,COMPSCI,['C']
+161,CA,"['A', 'B']"
+162,DANCE,['A']
+163,DATASCI,['C']
+164,DIGICOMM,['A']
+165,DIGIHUM,['B']
+166,DISABST,['A']
+167,EARTHSCI,['C']
+168,ECONOMIC,['A']
+169,EDUC,['A']
+170,ECE,['C']
+171,ELI,['C']
+172,ENGSCI,['C']
+173,ENGLISH,['B']
+174,ENVIRSCI,['C']
+175,EPID,['C']
+176,EPIDEMIO,['C']
+177,FIMS,['A']
+178,FAMLYSTU,['A']
+179,FILM,['B']
+180,FINMOD,['C']
+181,FOODNUTR,['A']
+182,FRENCH,['B']
+183,GSWS,"['A', 'B']"
+184,GEOGRAPH,['A']
+185,GERMAN,['B']
+186,GGB,['B']
+187,GLE,['A']
+188,GREEK,['B']
+189,GPE,['C']
+190,HEALTSCI,['A']
+191,HEBREW,['B']
+192,HINDI,['B']
+193,HISTORY,['A']
+194,HISTSCI,['C']
+195,HUMANECO,['A']
+196,HUMANRS,['A']
+197,HUMANIT,[]
+198,INDIGSTU,['A']
+199,IE,['C']
+200,INTEGSCI,['C']
+201,ICC,['B']
+202,INTERDIS,['A']
+203,INTREL,['A']
+204,ITALIAN,['B']
+205,ITALST,['B']
+206,JAPANESE,['B']
+207,JEWISH,['A']
+208,KINESIOL,['A']
+209,KOREAN,['B']
+210,LATIN,['B']
+211,LAW,['A']
+212,LS,['A']
+213,LINGUIST,"['A', 'B']"
+214,MOS,['A']
+215,MATH,['C']
+216,MME,['C']
+217,MSE,['C']
+218,MIT,['A']
+219,MBI,['C']
+220,MEDBIO,['C']
+221,MEDSCIEN,['C']
+222,MEDIEVAL,['B']
+223,MICROIMM,['C']
+224,MCS,['B']
+225,MUSIC,['A']
+226,NEURO,['C']
+227,NMM,['C']
+228,NURSING,['A']
+229,ONEHEALT,['C']
+230,PATHOL,['C']
+231,PERSIAN,['B']
+232,PHARM,['C']
+233,PHILOSOP,['B']
+234,PHYSICS,['C']
+235,PHYSIOL,['C']
+236,PHYSPHRM,['C']
+237,POLISCI,['A']
+238,PPE,['A']
+239,PORTUGSE,['B']
+240,PSYCHOL,['A']
+241,REHABSCI,['A']
+242,RELSTUD,['B']
+243,RUSSIAN,['B']
+244,SCHOLARS,[]
+245,SCIENCE,['C']
+246,SOCLJUST,['A']
+247,SOCSCI,['A']
+248,SOCWORK,['A']
+249,SOCIOLOG,['A']
+250,SE,['C']
+251,SPANISH,['B']
+252,SPEECH,['B']
+253,STATS,['C']
+254,SA,['B']
+255,THANAT,['A']
+256,TNLA,['B']
+257,THEATRE,['B']
+258,THEOLST,['B']
+259,TJ,['A']
+260,WORLDLIT,['B']
+261,WRITING,['B']
